Try to fix decompression

+++ app/src/main/java/com/kunzisoft/keepass/database/action/SaveDatabaseRunnable.kt
@@ -30,11 +30,13 @@ open class SaveDatabaseRunnable(protected var context: Context,
     : ActionRunnable() {
 
     var mAfterSaveDatabase: ((Result) -> Unit)? = null
+    protected var mAllowRemoveUnlinkedAttachments = true
 
     override fun onStartRun() {}
 
     override fun onActionRun() {
-        database.removeUnlinkedAttachment()
+        if (mAllowRemoveUnlinkedAttachments)
+            database.removeUnlinkedAttachment()
         if (saveDatabase && result.isSuccess) {
             try {
                 database.saveData(context.contentResolver)
